+++ src/api/api.go
@@ -2,6 +2,8 @@ package api
 
 import (
 	"github.com/dattito/purrmannplus-backend/api/providers"
+	"github.com/dattito/purrmannplus-backend/config"
+	"github.com/dattito/purrmannplus-backend/utils/logging"
 )
 
 var api providers.Provider
@@ -12,5 +14,9 @@ func Init() error {
 }
 
 func StartListening() error {
-	return api.StartListening()
+	logging.Infof("Starting listening on port %d", config.LISTENING_PORT)
+	if config.ENABLE_API {
+		return api.StartListening()
+	}
+	return nil
 }